# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPAY API",
        "description": "API pour le système de paiement mobile MOMPAY",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get all transactions",
                "description": "Display a listing of transactions.",
                "operationId": "6581469356772e7a74a0d79897544f1a",
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "montant": {
                                                        "type": "string",
                                                        "example": "+100.00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/transfer": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Perform a transfer",
                "description": "Perform a transfer transaction.",
                "operationId": "c52a2fa87fe3a4d506bb209f47d65b2e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero",
                                    "montant"
                                ],
                                "properties": {
                                    "numero": {
                                        "type": "string",
                                        "example": "771234567"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transfer completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/payment": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Perform a payment",
                "description": "Perform a payment transaction.",
                "operationId": "7f31f89f19b7010d709abbf8eadd1e25",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero",
                                    "montant"
                                ],
                                "properties": {
                                    "numero": {
                                        "description": "Numéro de téléphone ou code marchand",
                                        "type": "string"
                                    },
                                    "montant": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter your Bearer token"
            },
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter your Bearer token"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}